+++ index.html
@@ -25,11 +25,13 @@
             <div class="slotBoxes" id="11"></div>
         </div>
         <div class="btnContainer">
-            <button>SPIN THE REELS</button>
+            <button id="spinButton">SPIN THE REELS</button>
+            <div id="cashAmount"></div>
+            <div id="betLevel">Bet Level 1</div>
         </div>
     </div>
 
-
+<script src="./dist/main.js"></script>
 </body>
 
 </html>
